requirements.txt
SQLAlchemy==1.3.23
requests==2.25.1
pandas==1.2.2
PyYAML==5.4.1
xlsxwriter
